Fix the state labels in the data point circles
Remove the text-anchor and fill from circleLabels and replace it with the stateText class that has styling in d3Style.css.
Remove the fill from the circles selection and replace it with the stateCircle class that has styling in d3Style.css.
Move the circleLabels selection after the circles selection.