Modify and move the labels
Create a xLabelsGroup and a yLabelsGroup.
Add the aText class to both labels groups for styling.
Add a transform attribute and a font-weight attribute to both labels groups for alignment and styling, respectively.
Add a value attribute to each label to grab the corresponding data value for the label.
Remove the x and y attributes from the x-axis labels.
Remove the transform (rotate)  attribute from the y-axis labels.
Remove the text-anchor, font-size, and font-weight attributes/styles from all of the x-axis and y-axis labels.
Move all of the labels after circlesGroup is created.
Change circles to circlesGroup.
Remove all of the commented/incomplete x-axis updating functions.
Remove the commented/incomplete onclick event listener.