Fix the tooltips positions
Remove the tooltip update function and place some of the comparison 
operator code used to determine the label values inside .html().
Remove the update tooltip functions in the event listeners.
Add the mouse events to the update pattern section.
Call the tooltip after circlesGroup is created.
Replace the arrow functions in the mouse events with regular anonymous 
functions.